Moved all JSForce calls into ElectronForce.js
